[setup] Getting the initial setup ready
The platformio project has been changed to not include every log level (perhaps that might be the reason why the esp was disconnected on windows) and removed the button tests, because they were causing a ruckus on Windows. However, just the basic project with these changes works on the same PC with the same Bluetooth adapter only running Manjaro instead of Windows. I don't know what might be the issue, but until I can develop against the HID library and can test it on Linux, I'll just do that and will figure out what's the matter with Windows later.